requirements.txt
numpy
matplotlib
pandas
finviz
requests
orgasm-package
scikit-learn